// packages/cli/src/components/command-menu/types.ts
import type { DialogContextValue } from "../providers/dialog";
import type { ToastContextValue } from "../providers/toast";
import type { ModeType, SupportedChatModelId } from "@daycode/shared";

export type CommandContext = {
    exit: () => void;
    toast: ToastContextValue;
    dialog: DialogContextValue;
    navigate: (path: string) => void;
    mode: ModeType;
    setMode: (mode: ModeType) => void;
    setModel: (model: SupportedChatModelId)=> void;
}

export type Command = {
    name: string;
    description: string;
    value: string;
    action?: (ctx: CommandContext) => void | Promise<void> ;
}